take over #2273 from tjhei/split_instantiations
Split instantiations of matrix_tools.cc
Conflicts:
	source/numerics/matrix_tools.cc
	source/numerics/matrix_tools.inst.in
